feat: build and scan with images cleanup https://github.com/canonical/notebook-operators/issues/188
Summary of changes:
- Added script that builds and scans with removal of intermediate and
  scanned images to save space. This is due to limited disk space on
free Github runners.

+++ image-definitions/scan.sh
@@ -9,6 +9,14 @@
 set -e
 
 TAG=$1
+REPORT_TOTALS=false
+TRIVY_REPORTS_DIR="trivy-reports"
+TRIVY_REPORT_TYPE="json"
+
+if [ -d "$TRIVY_REPORTS_DIR" ]; then
+    echo "WARNING: $TRIVY_REPORTS_DIR directory already exists. Some reports might not be generated."
+    echo "         To scan all images remove $TRIVY_REPORTS_DIR directory."
+fi
 
 # Kubeflow container images scan
 echo "Scan container images for Kubeflow"
@@ -17,12 +25,18 @@ REPO_DIR="kubeflow"
 TAG=${TAG:-$(eval "cat $REPO_DIR/version.txt")}
 KF_PATCH_COMMIT=$(eval "cat ./kubeflow-patch-commit.txt")
 DATE=$(date +%F)
-SCAN_SUMMARY_FILE="scan-summary-$DATE.txt"
-echo "Tag: $TAG" > $SCAN_SUMMARY_FILE
-echo "Date: $DATE" >> $SCAN_SUMMARY_FILE
+SCAN_SUMMARY_FILE="scan-summary.txt"
+if [ ! -f $SCAN_SUMMARY_FILE ]; then
+    # create header for scan summary file, if it does not exist
+    echo "Tag: $TAG" > $SCAN_SUMMARY_FILE
+    echo "Date: $DATE" >> $SCAN_SUMMARY_FILE
+    echo "CVEs per image:" >> $SCAN_SUMMARY_FILE
+    echo " IMAGE | BASE | CRITICAL | HIGH | MEDIUM | LOW " >> $SCAN_SUMMARY_FILE
+    echo " -- | -- | -- | -- | -- | -- " >> $SCAN_SUMMARY_FILE
+fi
 
 # create directory for trivy reports
-mkdir -p trivy-reports
+mkdir -p "$TRIVY_REPORTS_DIR"
 
 # get all images that are available for scanning
 # excluded:
@@ -30,33 +44,38 @@ mkdir -p trivy-reports
 # - aquasec/trivy repository (scanner)
 IMAGE_LIST=($(docker images --format="{{json .}}" | jq -r 'select((.Tag=="$TAG") or (.Tag!="<none>" and .Tag!="$TAG-$KF_PATCH_COMMIT" and .Repository!="aquasec/trivy")) | "\(.Repository):\(.Tag)"'))
 
-echo "CVEs per image:" >> $SCAN_SUMMARY_FILE
-echo " IMAGE | BASE | CRITICAL | HIGH | MEDIUM | LOW " >> $SCAN_SUMMARY_FILE
-echo " -- | -- | -- | -- | -- | -- " >> $SCAN_SUMMARY_FILE
-# for every image generate trivy report and store it in `trivy-reports/` directory
+# for every image generate trivy report and store it in `$TRIVY_REPORTS_DIR/` directory
 # ':' and '/' in image names are replaced with '-' for files
 for IMAGE in "${IMAGE_LIST[@]}"; do
     # trivy report name should contain artifact name being scanned with ':' and '/' replaced with '-'
     TRIVY_REPORT="$IMAGE"
     TRIVY_REPORT=$(echo $TRIVY_REPORT | sed 's/:/-/g')
     TRIVY_REPORT=$(echo $TRIVY_REPORT | sed 's/\//-/g')
-    TRIVY_REPORT=$(echo "trivy-reports/$TRIVY_REPORT")
-    docker run -v /var/run/docker.sock:/var/run/docker.sock -v `pwd`:`pwd` -w `pwd` aquasec/trivy image -f json -o $TRIVY_REPORT.json --ignore-unfixed $IMAGE
-    NUM_CRITICAL=$(grep CRITICAL $TRIVY_REPORT.json | wc -l)
-    NUM_HIGH=$(grep HIGH $TRIVY_REPORT.json | wc -l)
-    NUM_MEDIUM=$(grep MEDIUM $TRIVY_REPORT.json | wc -l)
-    NUM_LOW=$(grep LOW $TRIVY_REPORT.json | wc -l)
-    BASE=$(cat $TRIVY_REPORT.json | jq '.Metadata.OS | "\(.Family):\(.Name)"' | sed 's/"//g')
-    echo " $IMAGE | $BASE | $NUM_CRITICAL | $NUM_HIGH | $NUM_MEDIUM | $NUM_LOW " >> $SCAN_SUMMARY_FILE
-    # remove image to save space
-    docker rmi -f $IMAGE
+    TRIVY_REPORT=$(echo "$TRIVY_REPORTS_DIR/$TRIVY_REPORT.$TRIVY_REPORT_TYPE")
+    if [ -f "$TRIVY_REPORT" ]; then
+      echo "Trivy report '$TRIVY_REPORT' for $IMAGE already exist, skip it"
+      continue
+    fi
+    echo "Scan image $IMAGE report in $TRIVY_REPORT"
+    docker run -v /var/run/docker.sock:/var/run/docker.sock -v `pwd`:`pwd` -w `pwd` aquasec/trivy image -f $TRIVY_REPORT_TYPE -o $TRIVY_REPORT --ignore-unfixed $IMAGE
+    if [ "$TRIVY_REPORT_TYPE" = "json" ]; then
+      # for JSON type retrieve severity counts
+      NUM_CRITICAL=$(grep CRITICAL $TRIVY_REPORT | wc -l)
+      NUM_HIGH=$(grep HIGH $TRIVY_REPORT | wc -l)
+      NUM_MEDIUM=$(grep MEDIUM $TRIVY_REPORT | wc -l)
+      NUM_LOW=$(grep LOW $TRIVY_REPORT | wc -l)
+      BASE=$(cat $TRIVY_REPORT | jq '.Metadata.OS | "\(.Family):\(.Name)"' | sed 's/"//g')
+      echo " $IMAGE | $BASE | $NUM_CRITICAL | $NUM_HIGH | $NUM_MEDIUM | $NUM_LOW " >> $SCAN_SUMMARY_FILE
+    fi
 done
 
-NUM_CRITICAL=$(grep CRITICAL trivy-reports/* | wc -l)
-NUM_HIGH=$(grep HIGH trivy-reports/* | wc -l)
-NUM_MEDIUM=$(grep MEDIUM trivy-reports/* | wc -l)
-NUM_LOW=$(grep LOW trivy-reports/* | wc -l)
-echo "| | Totals: | $NUM_CRITICAL | $NUM_HIGH | $NUM_MEDIUM | $NUM_LOW" >> $SCAN_SUMMARY_FILE
+if [ "$REPORT_TOTALS" = true ]; then
+    NUM_CRITICAL=$(grep CRITICAL "$TRIVY_REPORTS_DIR/*" | wc -l)
+    NUM_HIGH=$(grep HIGH "$TRIVY_REPORTS_DIR/*" | wc -l)
+    NUM_MEDIUM=$(grep MEDIUM "$TRIVY_REPORTS_DIR/*" | wc -l)
+    NUM_LOW=$(grep LOW "$TRIVY_REPORTS_DIR/*" | wc -l)
+    echo "| | Totals: | $NUM_CRITICAL | $NUM_HIGH | $NUM_MEDIUM | $NUM_LOW" >> $SCAN_SUMMARY_FILE
+fi
 cat $SCAN_SUMMARY_FILE
 
 # End of Kubeflow container images scan
